+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
-
+- added stat cards to the group view
+- added stat cards to the student view
 - Added a clear icon to search input and removed dead student performance view
 - Added a message to search results when no results are found